requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.0.1
chromedriver-binary==138.0.7158.0.0
click==8.0.3
colorama==0.4.4
e621py-wrapper==0.1.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.116.1
Flask==2.0.2
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
MouseInfo==0.1.3
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.53
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.2.3
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==1.1.1
pytweening==1.0.4
qrcode==8.2
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.32.0
simplejson==3.17.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==1.26.14
uvicorn==0.35.0
websocket-client==1.8.0
Werkzeug==2.0.2
wsproto==1.2.0
